handler addUserProfile

+++ profileService/internal/domain/userinstitute/entity.go
@@ -4,5 +4,5 @@ type UserInstitute struct {
 	UserInstituteID  int
 	IsRepresentative bool
 	InstituteID      int
-	ProfileID        int
+	ProfileID        int64
 }
